rag_agent/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.9
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
google-generativeai>=0.3.0
pypdf>=4.0.0
PyGithub>=2.1.1
streamlit>=1.30.0
chromadb>=0.4.22
tiktoken>=0.5.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
